refactor: context로직 분리
